fuse compositional
